refactor: move sampling classes to ai.vespa.sampling in vespajlib
