+++ container-search/src/test/java/com/yahoo/vespa/streamingvisitors/tracing/SamplingTraceExporterTest.java
@@ -1,6 +1,7 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.streamingvisitors.tracing;
 
+import ai.vespa.sampling.SamplingStrategy;
 import org.junit.jupiter.api.Test;
 
 import static org.mockito.ArgumentMatchers.any;
