+++ vespajlib/src/test/java/ai/vespa/sampling/MaxSamplesPerPeriodTest.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.streamingvisitors.tracing;
+package ai.vespa.sampling;
 
 import org.junit.jupiter.api.Test;
